// src/services/generated/models/tokenPairResponse.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * SEMD API
 * OpenAPI spec version: 1.0.0
 */

export interface TokenPairResponse {
  /** JWT access token */
  access_token: string;
  message?: string;
  /** JWT refresh token */
  refresh_token: string;
  status?: number;
  /** Token type */
  token_type?: string;
}