// src/services/generated/models/preAuthResponse.ts
/**
 * Generated by orval v8.5.1 🍺
 * Do not edit manually.
 * SEMD API
 * OpenAPI spec version: 1.0.0
 */

export interface PreAuthResponse {
  status?: number;
  message?: string;
  /** Pre-authentication token for 2FA */
  pre_auth_token: string;
  /** Indicates 2FA is required */
  requires_2fa?: boolean;
}